Ignore file sections
There are some supportconfig files sections which are grep commands in
some files in the file system. As these files will be present after the
split of all the supportconfig files we can ignore such files. Thus,
this commit allows the user of the go-supportconfig library to tell the
splitter which section should be ignored. For that, the user can return
a empty string in the path handler function. The go-supportconfig
library will skip that section.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>